event_mgr: move event management out of xenstore
The event manager is responsible for running callbacks and pool for io,
and run more callbacks, and poll for io...
There should be only two types of callbacks, generic and io related, but
current implementation is still aware of watches as a type of callback.
This should be fixed later on.

The main idea for the event manager is to simplify a multi-threaded
implementation later on. After adding locks where necessary the event
manager can start executing callbacks in multiple threads coming from a
thread pool.