*_conn: provide need_{tx,rx} methods
Before, the callbacks process_{rx,tx} would be called only when more IO
was requested by reading from or writing to the connection. These
methods indicate that the appropriate callback should be called as soon
as IO is possible, without the need to do IO.

These methods might be used by the protocol implementation in situations
where it is necessary to indicate that IO is needed, but the actual IO
can't be done in the current context, like in the ctor.