xs_proto_v1: move protocol impl out of client
A new namespace was created for the implementation of the upstream
version of the xenstore protocol (here called v1). This makes the
implementation of new (future) protocols easier by separating what is
specific to the protocol from the rest.

Arguably the client class isn't necessary anymore, given pretty much all
of its content was moved to protocol. However it will be kept for now,
given it can be used in the future to implement protocol independent
functionality.

The new implementation is very similar to the previous one in the client
class. However it brings the following improvements:

* Cleaner code
* Better error checking/handling
* Independent handling of tx and rx