client: make use of the new xs_proto_v1
This commit just removes pretty much all of the code of the client class
by replacing it with the new protocol implementation. It also makes the
necessary changes in depend classes.

In terms of logging, opening and closing connections don't depend on the
DEBUG flag any more. This should anyway be improved in the future by
adding a logging facility.