sock_conn: use MSG_NOSIGNAL to avoid SIGPIPE
Previously SIGPIPEs where being ignored by setting the signal handler to
SIG_IGN on app/lixs.cc. This was sort of an hack, given it depends on
the application setting it. With this commit, writes to the socket are
done with the MSG_NOSIGNAL flag, explicitly preventing SIGPIPEs all
together.

Fixes #1.