event_mgr: enable concurrent enqueue_event
The list iterator is not invalidated by a adding or removing elements,
as far as the change doesn't occur in the element currently referenced
by the iterator. Therefore it is possible to constantly feed the
event_list while the events are being dispatched.

This commit makes the necessary changes to ensure this behaviour works
correctly.

Fixes #5.