iomux: add error information to io callback
This commits adds a boolean to the io_cb operator() that indicates if a
error happened for that associated file descriptor. For this iteration
the information is just binary, later more detail information can be
added by replacing the boolean with an error type.

This commit doesn't include handling of the error by the iomux clients.

Fixes #10.