iomux: redefine API and re-implement clients
This commit provides a simpler API that dosn't require to keep an object
to execute operations. The callback is also a simple std::function to be
inline with the event_mgr. This also simplifies to later use the
event_mgr to dispatch events instead of calling the callbacks directly.

For *_conn the support for this new API was implemented resorting to
shared pointers, enabling the callback to verify if the source object
was already deleted, therefore this commit fixes #12.