Makefile: add basic configuration system
This basically lets you define configuration variables in a
configuration file instead of passing them to the make command line.
Currently the only configuration is CONFIG_LOGGER_MAX_LEVEL.

All the targets will be re-built every time the configuration file is
updated. This is not perfect but is more than enough for such a small
project.

Finally The target configure is provided to generate the default
configuration file. If necessary a different configuration file can be
provided in the command line by specifying `config`. Changing the
configuration file won't make all targets recompile, so don't forget to
`make clean` before.